add lld and lli option/flags for my printf

+++ libft/ft_putstr_fd.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-void		ft_putstr_fd(char *s, int fd)
+void			ft_putstr_fd(char *s, int fd)
 {
-	size_t	i;
+	long long	i;
 
 	i = 0;
 	if (s != NULL)
